Fase 2/Evidencias Proyecto/Desarrollo del proyecto/Version Nueva del proyecto/requirements.txt
django-quill-editor
django-crispy-forms
django-livereload-server
